Extract tariff models into separate files
